SAAS-283 componentize stuff
