+++ packages/platform-ui/src/components/Anonymous/Anonymous.messages.ts
@@ -1,0 +1,6 @@
+export const Messages = {
+  companyName: 'Percona',
+  description:
+    'Single pane of glass for managing and monitoring the performance of your MySQL, MariaDB, PostgreSQLand MongoDB databases.',
+  logoAlt: 'Percona Logo',
+};
