+++ packages/platform-ui/src/components/Anonymous/index.ts
@@ -1,0 +1,1 @@
+export * from './Anonymous';
