+++ packages/platform-ui/src/components/Root/index.ts
@@ -1,0 +1,2 @@
+export * from './Root.component';
+export * from './Root.route';
