+++ packages/platform-ui/src/components/TestContainer/index.ts
@@ -1,0 +1,1 @@
+export * from './TestContainer';
